+++ src/main/java/roomescape/controller/dto/reservationtime/AvailableReservationTimeResponse.java
@@ -2,7 +2,7 @@ package roomescape.controller.dto.reservationtime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.service.dto.AvailableReservationTimeResult;
+import roomescape.service.dto.reservationtime.AvailableReservationTimeResult;
 
 public record AvailableReservationTimeResponse(
         Long id,
