+++ src/main/java/roomescape/service/dto/reservation/ReservationPagingCondition.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.reservation;
 
 import roomescape.global.exception.InvalidReservationPagingException;
