+++ src/main/java/roomescape/service/dto/reservationtime/AvailableReservationTimesResult.java
@@ -1,6 +1,7 @@
-package roomescape.service.dto;
+package roomescape.service.dto.reservationtime;
 
 import java.util.List;
+import roomescape.service.dto.theme.ThemeResult;
 
 public record AvailableReservationTimesResult(
         ThemeResult theme,
