+++ src/main/java/roomescape/service/dto/reservationtime/CreateReservationTimeCommand.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.reservationtime;
 
 import java.time.LocalTime;
 import roomescape.global.exception.InvalidReservationTimeException;
