+++ src/main/java/roomescape/service/dto/reservationtime/ReservationTimeResult.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.reservationtime;
 
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
