+++ src/main/java/roomescape/service/dto/theme/ThemeRankingCondition.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.theme;
 
 import roomescape.global.exception.InvalidRankingConditionException;
